+++ pages/results.py
@@ -1,0 +1,9 @@
+import dash
+from dash import html
+
+dash.register_page(__name__)
+
+layout = html.Div([
+    html.H1('This is our Archive page'),
+    html.Div('This is our Archive page content.'),
+])
